requirement.txt
Django==2.2.12
django-extensions==2.2.9
django-summernote==0.8.11.6
django-taggit==1.3.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
